refactor(core): extract complete relay capacity policy
