+++ packages/core/src/index.ts
@@ -128,9 +128,6 @@ export {
 export * from './crypto/index.js';
 export * from './proto/index.js';
 export {
-  DKGNode,
-  type RelayStats,
-  type RelayReservationDetail,
   // Capacity helpers + constants from PR #524 (libp2p reachability hardening PR1).
   // Re-exported here so dashboards / route handlers can import them
   // without reaching into the deep import path.
@@ -148,6 +145,11 @@ export {
   MAX_RELAY_RESERVATION_COUNT,
   validateRelayReservationCount,
   type RelayReservationCountValidation,
+} from './relay-capacity-policy.js';
+export {
+  DKGNode,
+  type RelayStats,
+  type RelayReservationDetail,
   // Single-source-of-truth interface for the small / sparse-network
   // tunables forwarded `DkgConfig.network` → `DKGAgentConfig` →
   // `DKGNodeConfig`, plus the choke-point helper every forwarder
