+++ packages/core/test/relay-capacity.test.ts
@@ -196,12 +196,12 @@ describe('checkFdLimit', () => {
     }
   });
 
-  it('uses the relay capacity policy above the minimum FD floor', () => {
+  it('applies explicit FD headroom above the minimum floor', () => {
     const capacity = 8192;
     const emissions: string[] = [];
     checkFdLimit(capacity, (_level, message) => emissions.push(message), () => 4096);
     expect(emissions).toHaveLength(1);
-    expect(emissions[0]).toContain(`recommended ${deriveRelayCaps(capacity).maxConnections}`);
+    expect(emissions[0]).toContain('recommended 16384');
   });
 
   let readLimit: () => number | undefined;
